+++ jgiven-core/src/main/java/com/tngtech/jgiven/annotation/Hidden.java
@@ -7,16 +7,17 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotates a step method to be hidden in reports. 
- * 
- * This is useful for technical helper methods that have no meaning in the domain world.
- * 
- * You should write helper methods in camelCase so that the name already indicates that
+ * Annotates elements that should be hidden in reports. 
+ * <p>
+ * This is useful for technical helper methods or arguments that have no meaning for domain expert,
+ * but are needed for some technical reasons.
+ * <p>
+ * You should write technical helper methods in camelCase so that the name already indicates that
  * the method does not appear in the report.
  */
 @Documented
 @Retention( RetentionPolicy.RUNTIME )
-@Target( ElementType.METHOD )
+@Target( { ElementType.METHOD, ElementType.PARAMETER } )
 public @interface Hidden {
 
 }